pyproject.toml
[project]
name = "youtube-summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "python-dotenv>=1.1.0",
    "youtube-transcript-api>=1.2.2",
    "yt-dlp>=2025.7.21",
    "langchain-community>=0.3.7",
    "chromadb>=0.5.5",
    "langchain-groq>=0.3.7",
    "loguru>=0.7.3",
    "langchain-chroma>=0.2.5",
    "fastembed>=0.7.1",
]

[dependency-groups]
dev = ["pre-commit>=4.2.0", "ruff>=0.11.5", "pytest>=8.3.2", "mypy>=1.17.1"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D213"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.mypy]
strict = true
